Added DeepView UI
Showing all instances and their dependencies. This is for
debugging purposes, so needs to be enabled via init.lua
