+++ Components/WinFormUI/Forms/Helpers/IEventFilter.cs
@@ -1,0 +1,11 @@
+﻿#nullable enable
+
+using Slipstream.Shared;
+
+namespace Slipstream.Components.WinFormUI.Forms.Helpers
+{
+    internal interface IEventFilter
+    {
+        bool Accept(IEvent e);
+    }
+}
