+++ Components/WinFormUI/Forms/Helpers/NodeTypeEnum.cs
@@ -1,0 +1,13 @@
+﻿#nullable enable
+
+namespace Slipstream.Components.WinFormUI.Forms.Helpers
+{
+    internal enum NodeTypeEnum
+    {
+        None,
+        Generic,
+        LuaScripts,
+        Instance,
+        Dependency
+    }
+}
