+++ Components/WinFormUI/Forms/Helpers/NoneEventFilter.cs
@@ -1,0 +1,14 @@
+﻿#nullable enable
+
+using Slipstream.Shared;
+
+namespace Slipstream.Components.WinFormUI.Forms.Helpers
+{
+    internal class NoneEventFilter : IEventFilter
+    {
+        public bool Accept(IEvent e)
+        {
+            return false;
+        }
+    }
+}
